+++ src/main/java/roomescape/theme/service/ThemeService.java
@@ -1,9 +1,7 @@
 package roomescape.theme.service;
 
-import java.time.LocalDate;
 import java.util.List;
 
-import roomescape.time.domain.ReservationTime;
 import roomescape.theme.domain.Theme;
 import roomescape.theme.service.dto.ThemeSaveServiceDto;
 
@@ -11,7 +9,5 @@ public interface ThemeService {
     List<Theme> getAll();
     Theme create(ThemeSaveServiceDto theme);
     void deleteById(Long id);
-    List<ReservationTime> getAvailableTimes(Long themeId, LocalDate date);
     List<Theme> getBestThemes();
-
 }
