+++ src/test/java/roomescape/theme/service/ThemeServiceImplTest.java
@@ -4,7 +4,6 @@ import java.time.Clock;
 import java.time.Instant;
 import java.time.LocalDate;
 import java.time.ZoneId;
-import java.util.Collections;
 import java.util.List;
 
 import static org.assertj.core.api.Assertions.assertThat;
@@ -16,34 +15,20 @@ import org.mockito.ArgumentCaptor;
 import static org.mockito.ArgumentMatchers.any;
 import org.mockito.Mock;
 import static org.mockito.Mockito.verify;
-import static org.mockito.Mockito.verifyNoInteractions;
 import static org.mockito.Mockito.when;
 import org.mockito.junit.jupiter.MockitoExtension;
 
-import roomescape.holiday.repository.HolidayRepository;
-import roomescape.time.domain.ReservationTime;
-import roomescape.reservation.repository.ReservationRepository;
 import roomescape.theme.domain.Theme;
 import roomescape.theme.exception.ThemeNotFoundException;
 import roomescape.theme.repository.ThemeRepository;
 import roomescape.theme.service.dto.ThemeBestServiceDto;
 import roomescape.theme.service.dto.ThemeSaveServiceDto;
-import roomescape.time.service.TimeService;
 
 @ExtendWith(MockitoExtension.class)
 class ThemeServiceImplTest {
 
     @Mock
     private ThemeRepository themeRepository;
-
-    @Mock
-    private TimeService timeService;
-
-    @Mock
-    private HolidayRepository holidayRepository;
-
-    @Mock
-    private ReservationRepository reservationRepository;
 
     private ThemeServiceImpl themeService;
     private Clock clock;
@@ -53,9 +38,6 @@ class ThemeServiceImplTest {
         clock = Clock.fixed(Instant.parse("2026-05-06T00:00:00Z"), ZoneId.of("Asia/Seoul"));
         themeService = new ThemeServiceImpl(
                 themeRepository,
-                timeService,
-                holidayRepository,
-                reservationRepository,
                 clock,
                 7,
                 10
@@ -110,67 +92,6 @@ class ThemeServiceImplTest {
                 .hasMessage("테마가 존재하지 않습니다. id=99")
                 .extracting(e -> ((ThemeNotFoundException) e).getId())
                 .isEqualTo(99L);
-    }
-
-    @Test
-    void getAvailableTimes_테마가없으면_예외() {
-        LocalDate date = LocalDate.of(2026, 5, 6);
-        when(themeRepository.existsById(999L)).thenReturn(false);
-
-        assertThatThrownBy(() -> themeService.getAvailableTimes(999L, date))
-                .isInstanceOf(ThemeNotFoundException.class)
-                .hasMessage("테마가 존재하지 않습니다. id=999");
-
-        verifyNoInteractions(timeService, holidayRepository, reservationRepository);
-    }
-
-    @Test
-    void getAvailableTimes_휴일이면_빈리스트() {
-        LocalDate date = LocalDate.of(2026, 5, 6);
-        when(themeRepository.existsById(1L)).thenReturn(true);
-        when(holidayRepository.existsByDate(date)).thenReturn(true);
-
-        assertThat(themeService.getAvailableTimes(1L, date)).isEmpty();
-
-        verify(holidayRepository).existsByDate(date);
-        verifyNoInteractions(timeService, reservationRepository);
-    }
-
-    @Test
-    void getAvailableTimes_예약된시간은_제외한다() {
-        LocalDate date = LocalDate.of(2026, 5, 6);
-        when(themeRepository.existsById(1L)).thenReturn(true);
-        when(holidayRepository.existsByDate(date)).thenReturn(false);
-        when(reservationRepository.findTimeIdsByThemeIdAndDate(1L, date)).thenReturn(List.of(2L));
-
-        List<ReservationTime> allTimes = List.of(
-                new ReservationTime(1L, "10:00", "12:00"),
-                new ReservationTime(2L, "12:00", "14:00"),
-                new ReservationTime(3L, "14:00", "16:00")
-        );
-        when(timeService.findAll()).thenReturn(allTimes);
-
-        List<ReservationTime> results = themeService.getAvailableTimes(1L, date);
-
-        assertThat(results).extracting(ReservationTime::getId)
-                .containsExactly(1L, 3L);
-    }
-
-    @Test
-    void getAvailableTimes_예약이없으면_전체시간을_반환한다() {
-        LocalDate date = LocalDate.of(2026, 5, 6);
-        when(themeRepository.existsById(1L)).thenReturn(true);
-        when(holidayRepository.existsByDate(date)).thenReturn(false);
-        when(reservationRepository.findTimeIdsByThemeIdAndDate(1L, date)).thenReturn(Collections.emptyList());
-
-        List<ReservationTime> allTimes = List.of(
-                new ReservationTime(1L, "10:00", "12:00"),
-                new ReservationTime(2L, "12:00", "14:00")
-        );
-        when(timeService.findAll()).thenReturn(allTimes);
-
-        assertThat(themeService.getAvailableTimes(1L, date))
-                .isEqualTo(allTimes);
     }
 
     @Test
